refactor: HTTPMethod Enum 추가 및 패키지 리팩토링
