+++ src/main/java/webserver/http/domain/RequestParams.java
@@ -1,4 +1,4 @@
-package webserver.http;
+package webserver.http.domain;
 
 import java.util.Arrays;
 import java.util.HashMap;
