+++ src/test/java/webserver/http/PathTest.java
@@ -2,6 +2,8 @@ package webserver.http;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import webserver.http.domain.Path;
+import webserver.http.domain.RequestParams;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static webserver.http.RequestTestConstant.*;
